feat: 형 변환 손실 경고 메세지 해결 및 스테이지 플로팅 메세지 구현

+++ AvoidGame/Game/Actor/Enemy/EnemyA.cpp
@@ -38,7 +38,7 @@ EnemyA::EnemyA(const char* image, float speed)
 	}
 
 	// 초기 위치 설정
-	SetPosition(Vector2(xPosition, yPosition));
+	SetPosition(Vector2(static_cast<int>(xPosition), static_cast<int>(yPosition)));
 }
 
 void EnemyA::Update(float deltaTime)
