+++ AvoidGame/Game/Actor/Enemy/EnemyC.cpp
@@ -36,7 +36,7 @@ EnemyC::EnemyC(const char* image, float speed)
     }
 
     // 초기 위치 설정
-    SetPosition(Vector2(xPosition, yPosition));
+    SetPosition(Vector2(static_cast<int>(xPosition), static_cast<int>(yPosition)));
 }
 
 void EnemyC::Update(float deltaTime)
